Refactor TablesPage component; replace CreateTable with TablesList and enhance table management features
