+++ apps/console/src/components/project/database/tables/components/index.ts
@@ -1,1 +1,1 @@
-export * from "./_create_table";
+export * from "./_list";
