+++ apps/console/src/data/table-editor/table-editor-types.ts
@@ -79,3 +79,36 @@ export function isMaterializedView(entity?: Entity): entity is MaterializedView 
 export function isViewLike(entity?: Entity): entity is View | MaterializedView {
   return isView(entity) || isMaterializedView(entity);
 }
+
+export function postgresTableToEntity(table: PostgresTable): Entity | undefined {
+  if (table.columns === undefined || table.relationships === undefined) {
+    console.error(
+      'Unable to convert PostgresTable to Entity type: columns and relationships must not be undefined.'
+    )
+    return undefined
+  }
+
+  const tableRelationships: TableRelationship[] = table.relationships.map((rel) => ({
+    deletion_action: 'a',
+    update_action: 'a',
+    ...rel,
+  }))
+
+  return {
+    id: table.id,
+    schema: table.schema,
+    name: table.name,
+    comment: table.comment,
+    rls_enabled: table.rls_enabled,
+    rls_forced: table.rls_forced,
+    replica_identity: table.replica_identity,
+    bytes: table.bytes,
+    size: table.size,
+    live_rows_estimate: table.live_rows_estimate,
+    dead_rows_estimate: table.dead_rows_estimate,
+    columns: table.columns,
+    relationships: tableRelationships,
+    primary_keys: table.primary_keys,
+    entity_type: ENTITY_TYPE.TABLE,
+  }
+}
